Push helpers to custom TestCase class
