+++ composer.json
@@ -1,7 +1,8 @@
 {
   "autoload": {
     "psr-4": {
-      "TrendingRepos\\": "src/"
+      "TrendingRepos\\": "src/",
+      "Tests\\": "tests"
     }
   },
   "require": {
