+++ tests/RouterTest.php
@@ -1,9 +1,10 @@
 <?php
 
-use PHPUnit\Framework\TestCase;
+// use PHPUnit\Framework\TestCase;
 use TrendingRepos\Core\Router;
 use TrendingRepos\Core\Request;
 use TrendingRepos\Exception\RouterException;
+use Tests\TestCase;
 
 class RouterTest extends TestCase
 {
@@ -20,36 +21,18 @@ class RouterTest extends TestCase
     {
         $this->expectException(RouterException::class);
         $this->get('/wrong-url');
-        $this->router->get();
     }
 
     public function test_catch_exception_when_method_in_controller_doesnt_exist()
     {
         $this->expectException(RouterException::class);
         $this->get('/test');
-        $this->router->get();
     }
 
     public function test_catch_exception_when_routes_file_is_in_wrong_format()
     {
         $this->expectException(RouterException::class);
-        $this->get('/');
         $this->router = new Router(['/' => 'HomeController@home'], new Request);
-        $this->router->get();
-    }
-
-    protected function get(string $uri)
-    {
-        $this->call($uri, 'GET');
-    }
-
-    protected function post(string $uri)
-    {
-        $this->call($uri, 'POST');
-    }
-
-    private function call(string $uri, string $methodType) {
-        $_SERVER['PATH_INFO'] = $uri;
-        $_SERVER['REQUEST_METHOD'] = $methodType;
+        $this->get('/');
     }
 }
